cloud9: created
Cloud9 is a web-browser based IDE running on node.js.

This application should support multiple architectures as long as they have
node.js and a build of node-o3-xml that is dropped into this system.
Currently, that includes x86 32/64-bit and ARM armv7.  Ideally, that library
would be added separately as a system-dependent portion, because everything
else should be ISA independent.

Signed-off-by: Jason Kridner <jdk@ti.com>